Third Commit--a commit to the fix-issues branch...the second issue i.e the probelm of allowing anyone to edit or delete the images in the application...solution provide.....implement change in ImageController,ImageService,ImageRepository files and images.html file....

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId,HttpSession session, Model model) {
+        User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        boolean ownerOfImage = imageService.checkOwnerOfImage(user.getId(),imageId);
+        if (ownerOfImage) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }else{
+            String editError ="Only the owner of the image can edit the image";
+            model.addAttribute("image",image);
+            model.addAttribute("tags",image.getTags());
+            model.addAttribute("editError",editError);
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +149,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        boolean ownerOfImage = imageService.checkOwnerOfImage(user.getId(),imageId);
+        if (ownerOfImage) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else{
+            String deleteError ="Only the owner of the image can delete the image";
+            model.addAttribute("image",image);
+            model.addAttribute("tags",image.getTags());
+            model.addAttribute("deleteError",deleteError);
+            return "images/image";
+        }
+
     }
